+++ src/imio/pm/ws/WS4PM_types.py
@@ -108,7 +108,7 @@ class ns0:
         type = (schema, "ItemInfosRequest")
         def __init__(self, pname, ofwhat=(), attributes=None, extend=False, restrict=False, **kw):
             ns = ns0.ItemInfosRequest_Def.schema
-            TClist = [ZSI.TC.String(pname="UID", aname="_UID", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showExtraInfos", aname="_showExtraInfos", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showAnnexes", aname="_showAnnexes", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="include_annex_binary", aname="_include_annex_binary", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), GTD("http://ws4pm.imio.be","List",lazy=False)(pname="allowed_annexes_types", aname="_allowed_annexes_types", minOccurs=0, maxOccurs="unbounded", nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showAssembly", aname="_showAssembly", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showTemplates", aname="_showTemplates", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="inTheNameOf", aname="_inTheNameOf", minOccurs=0, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded"))]
+            TClist = [ZSI.TC.String(pname="UID", aname="_UID", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showExtraInfos", aname="_showExtraInfos", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showAnnexes", aname="_showAnnexes", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="include_annex_binary", aname="_include_annex_binary", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), GTD("http://ws4pm.imio.be","List",lazy=False)(pname="allowed_annexes_types", aname="_allowed_annexes_types", minOccurs=0, maxOccurs="unbounded", nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showAssembly", aname="_showAssembly", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showTemplates", aname="_showTemplates", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showEmptyValues", aname="_showEmptyValues", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="inTheNameOf", aname="_inTheNameOf", minOccurs=0, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded"))]
             self.attribute_typecode_dict = attributes or {}
             if extend: TClist += ofwhat
             if restrict: TClist = ofwhat
@@ -125,9 +125,38 @@ class ns0:
                     self._allowed_annexes_types = []
                     self._showAssembly = 0
                     self._showTemplates = 0
+                    self._showEmptyValues = 1
                     self._inTheNameOf = None
                     return
             Holder.__name__ = "ItemInfosRequest_Holder"
+            self.pyclass = Holder
+
+    class SingleItemInfosRequest_Def(ZSI.TCcompound.ComplexType, TypeDefinition):
+        schema = "http://ws4pm.imio.be"
+        type = (schema, "SingleItemInfosRequest")
+        def __init__(self, pname, ofwhat=(), attributes=None, extend=False, restrict=False, **kw):
+            ns = ns0.SingleItemInfosRequest_Def.schema
+            TClist = [ZSI.TC.String(pname="UID", aname="_UID", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showExtraInfos", aname="_showExtraInfos", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showAnnexes", aname="_showAnnexes", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="include_annex_binary", aname="_include_annex_binary", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), GTD("http://ws4pm.imio.be","List",lazy=False)(pname="allowed_annexes_types", aname="_allowed_annexes_types", minOccurs=0, maxOccurs="unbounded", nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showAssembly", aname="_showAssembly", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showTemplates", aname="_showTemplates", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.Boolean(pname="showEmptyValues", aname="_showEmptyValues", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="inTheNameOf", aname="_inTheNameOf", minOccurs=0, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded"))]
+            self.attribute_typecode_dict = attributes or {}
+            if extend: TClist += ofwhat
+            if restrict: TClist = ofwhat
+            ZSI.TCcompound.ComplexType.__init__(self, None, TClist, pname=pname, inorder=0, **kw)
+            class Holder:
+                __metaclass__ = pyclass_type
+                typecode = self
+                def __init__(self):
+                    # pyclass
+                    self._UID = None
+                    self._showExtraInfos = 0
+                    self._showAnnexes = 0
+                    self._include_annex_binary = None
+                    self._allowed_annexes_types = []
+                    self._showAssembly = 0
+                    self._showTemplates = 0
+                    self._showEmptyValues = 1
+                    self._inTheNameOf = None
+                    return
+            Holder.__name__ = "SingleItemInfosRequest_Holder"
             self.pyclass = Holder
 
     class ItemTemplateRequest_Def(ZSI.TCcompound.ComplexType, TypeDefinition):
@@ -650,6 +679,60 @@ class ns0:
                     self._itemInfo = []
                     return
             Holder.__name__ = "getItemInfosResponse_Holder"
+            self.pyclass = Holder
+
+    class getSingleItemInfosRequest_Dec(ElementDeclaration):
+        literal = "getSingleItemInfosRequest"
+        schema = "http://ws4pm.imio.be"
+        substitutionGroup = None
+        def __init__(self, **kw):
+            kw["pname"] = (u'http://ws4pm.imio.be', u'getSingleItemInfosRequest')
+            kw["aname"] = "_getSingleItemInfosRequest"
+            if ns0.SingleItemInfosRequest_Def not in ns0.getSingleItemInfosRequest_Dec.__bases__:
+                bases = list(ns0.getSingleItemInfosRequest_Dec.__bases__)
+                bases.insert(0, ns0.SingleItemInfosRequest_Def)
+                ns0.getSingleItemInfosRequest_Dec.__bases__ = tuple(bases)
+
+            ns0.SingleItemInfosRequest_Def.__init__(self, **kw)
+            if self.pyclass is not None: self.pyclass.__name__ = "getSingleItemInfosRequest_Dec_Holder"
+
+    class getSingleItemInfosResponse_Dec(ZSI.TCcompound.ComplexType, ElementDeclaration):
+        literal = "getSingleItemInfosResponse"
+        schema = "http://ws4pm.imio.be"
+        def __init__(self, **kw):
+            ns = ns0.getSingleItemInfosResponse_Dec.schema
+            TClist = [ZSI.TC.String(pname="UID", aname="_UID", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="id", aname="_id", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="title", aname="_title", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="creator", aname="_creator", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TCtimes.gDateTime(pname="creation_date", aname="_creation_date", minOccurs=0, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TCtimes.gDateTime(pname="modification_date", aname="_modification_date", minOccurs=0, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="category", aname="_category", minOccurs=0, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="description", aname="_description", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="detailedDescription", aname="_detailedDescription", minOccurs=0, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="decision", aname="_decision", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="preferredMeeting", aname="_preferredMeeting", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TCtimes.gDateTime(pname="preferred_meeting_date", aname="_preferred_meeting_date", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="review_state", aname="_review_state", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="meeting", aname="_meeting", minOccurs=0, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TCtimes.gDateTime(pname="meeting_date", aname="_meeting_date", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="item_assembly", aname="_item_assembly", minOccurs=0, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="absolute_url", aname="_absolute_url", minOccurs=1, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), ZSI.TC.String(pname="externalIdentifier", aname="_externalIdentifier", minOccurs=0, maxOccurs=1, nillable=False, typed=False, encoded=kw.get("encoded")), GTD("http://ws4pm.imio.be","AnnexInfo",lazy=False)(pname="annexes", aname="_annexes", minOccurs=0, maxOccurs="unbounded", nillable=False, typed=False, encoded=kw.get("encoded")), GTD("http://ws4pm.imio.be","TemplateInfo",lazy=False)(pname="templates", aname="_templates", minOccurs=0, maxOccurs="unbounded", nillable=False, typed=False, encoded=kw.get("encoded"))]
+            kw["pname"] = (u'http://ws4pm.imio.be', u'getSingleItemInfosResponse')
+            kw["aname"] = "_getSingleItemInfosResponse"
+            self.attribute_typecode_dict = {}
+            ZSI.TCcompound.ComplexType.__init__(self,None,TClist,inorder=0,**kw)
+            class Holder:
+                __metaclass__ = pyclass_type
+                typecode = self
+                def __init__(self):
+                    # pyclass
+                    self._UID = None
+                    self._id = None
+                    self._title = None
+                    self._creator = None
+                    self._creation_date = None
+                    self._modification_date = None
+                    self._category = None
+                    self._description = None
+                    self._detailedDescription = None
+                    self._decision = None
+                    self._preferredMeeting = None
+                    self._preferred_meeting_date = None
+                    self._review_state = None
+                    self._meeting = None
+                    self._meeting_date = None
+                    self._item_assembly = None
+                    self._absolute_url = None
+                    self._externalIdentifier = None
+                    self._annexes = []
+                    self._templates = []
+                    return
+            Holder.__name__ = "getSingleItemInfosResponse_Holder"
             self.pyclass = Holder
 
     class getItemTemplateRequest_Dec(ElementDeclaration):